[refactor] Create MatchingDiffUtil and MatchingViewHolder #164
