+++ app/src/main/java/find/ui/ui/matching/DibsAdapter.kt
@@ -2,17 +2,14 @@ package find.ui.ui.matching
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
-import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.ListAdapter
-import androidx.recyclerview.widget.RecyclerView
-import find.ui.BR
 import find.ui.databinding.ItemDibsBinding
 
 class DibsAdapter(private val isMyDibs: Boolean) :
-    ListAdapter<ResponseMatching, DibsAdapter.DibsViewHolder>(SendDibsDiffUtil()) {
+    ListAdapter<ResponseMatching, MatchingViewHolder<ItemDibsBinding>>(MatchingDiffUtil()) {
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int) =
-        DibsViewHolder(
+        MatchingViewHolder(
             ItemDibsBinding.inflate(
                 LayoutInflater.from(parent.context),
                 parent,
@@ -20,22 +17,6 @@ class DibsAdapter(private val isMyDibs: Boolean) :
             )
         )
 
-    override fun onBindViewHolder(holder: DibsViewHolder, position: Int) =
-        holder.bind(getItem(position))
-
-    inner class DibsViewHolder(private val binding: ItemDibsBinding) :
-        RecyclerView.ViewHolder(binding.root) {
-        fun bind(dibs: ResponseMatching) {
-            binding.setVariable(BR.dibs, dibs)
-            binding.setVariable(BR.isMyDibs, isMyDibs)
-        }
-    }
-
-    private class SendDibsDiffUtil : DiffUtil.ItemCallback<ResponseMatching>() {
-        override fun areItemsTheSame(oldItem: ResponseMatching, newItem: ResponseMatching) =
-            oldItem.userSequence == newItem.userSequence
-
-        override fun areContentsTheSame(oldItem: ResponseMatching, newItem: ResponseMatching) =
-            oldItem == newItem
-    }
+    override fun onBindViewHolder(holder: MatchingViewHolder<ItemDibsBinding>, position: Int) =
+        holder.dibsBind(getItem(position), isMyDibs)
 }
