+++ app/src/main/java/find/ui/ui/matching/FeelingAdapter.kt
@@ -2,17 +2,14 @@ package find.ui.ui.matching
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
-import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.ListAdapter
-import androidx.recyclerview.widget.RecyclerView
-import find.ui.BR
 import find.ui.databinding.ItemFeelingSendGetBinding
 
 class FeelingAdapter(private val isSend: Boolean) :
-    ListAdapter<ResponseMatching, FeelingAdapter.FeelingViewHolder>(SendFeelingDiffUtil()) {
+    ListAdapter<ResponseMatching, MatchingViewHolder<ItemFeelingSendGetBinding>>(MatchingDiffUtil()) {
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int) =
-        FeelingViewHolder(
+        MatchingViewHolder<ItemFeelingSendGetBinding>(
             ItemFeelingSendGetBinding.inflate(
                 LayoutInflater.from(parent.context),
                 parent,
@@ -20,22 +17,9 @@ class FeelingAdapter(private val isSend: Boolean) :
             )
         )
 
-    override fun onBindViewHolder(holder: FeelingViewHolder, position: Int) =
-        holder.bind(getItem(position))
-
-    inner class FeelingViewHolder(private val binding: ItemFeelingSendGetBinding) :
-        RecyclerView.ViewHolder(binding.root) {
-        fun bind(feeling: ResponseMatching) {
-            binding.setVariable(BR.feeling, feeling)
-            binding.setVariable(BR.isSend, isSend)
-        }
-    }
-
-    private class SendFeelingDiffUtil : DiffUtil.ItemCallback<ResponseMatching>() {
-        override fun areItemsTheSame(oldItem: ResponseMatching, newItem: ResponseMatching) =
-            oldItem.userSequence == newItem.userSequence
-
-        override fun areContentsTheSame(oldItem: ResponseMatching, newItem: ResponseMatching) =
-            oldItem == newItem
-    }
+    override fun onBindViewHolder(
+        holder: MatchingViewHolder<ItemFeelingSendGetBinding>,
+        position: Int
+    ) =
+        holder.feelingBind(getItem(position), isSend)
 }
